Upscale crops with kernelnet-guided anistropic kernel
Reference: arXiv:2112.09318v1

For unsharp, we use anistropic guided edge gating, so sharpening is not
uniform in the output.

+++ app/src/main/assets/shaders/CaptureSharpening/capturesharpening.glsl
@@ -1,5 +1,5 @@
-precision mediump float;
-precision mediump sampler2D;
+precision highp float;
+precision highp sampler2D;
 uniform sampler2D InputBuffer;
 uniform float size;
 uniform float strength;
